player ticker

+++ src/main/java/com/theendercore/eliptic/init/ELPConfig.java
@@ -8,7 +8,7 @@ public class ELPConfig {
     public static final ForgeConfigSpec SPEC;
     public static final ForgeConfigSpec.ConfigValue<Double> TP_HEIGHT;
     public static final ForgeConfigSpec.ConfigValue<Double> SINK_SPEED;
-    public static final ForgeConfigSpec.ConfigValue<Double> ECHO_TRIDENT_COOLDOWN;
+    public static final ForgeConfigSpec.ConfigValue<Integer> ECHO_TRIDENT_COOLDOWN;
 
     public ELPConfig() {
     }
@@ -18,7 +18,7 @@ public class ELPConfig {
         BUILDER.push("Config");
         TP_HEIGHT = BUILDER.comment("The height in blocks when the player uses the echo trident").define("tp_height", 15.0);
         SINK_SPEED = BUILDER.comment("the speed of the player sinking when using the echo trident").define("sink_speed", -0.05);
-        ECHO_TRIDENT_COOLDOWN = BUILDER.comment("the cooldown of the trident in ticks").define("echo_trident_cooldown", 200.0);
+        ECHO_TRIDENT_COOLDOWN = BUILDER.comment("the cooldown of the trident in ticks").define("echo_trident_cooldown", 200);
         BUILDER.pop();
         SPEC = BUILDER.build();
     }
